Refactoring for WireGuard support
