+++ pkg/cable/driver.go
@@ -1,7 +1,14 @@
-package ipsec
+package cable
 
 import (
+	"fmt"
+	"github.com/submariner-io/submariner/pkg/cable/ipsec"
 	"github.com/submariner-io/submariner/pkg/types"
+)
+
+const (
+	IPSec 		= "ipsec"
+	WireGuard 	= "wg"
 )
 
 // Driver is used by the ipsec engine to actually connect the tunnels.
@@ -19,4 +26,21 @@ type Driver interface {
 
 	// DisconnectFromEndpoint disconnects from the connection to the given endpoint.
 	DisconnectFromEndpoint(endpoint types.SubmarinerEndpoint) error
+}
+
+func NewDriver(localSubnets []string, localEndpoint types.SubmarinerEndpoint) (Driver, error) {
+	switch localEndpoint.Spec.Backend {
+	case IPSec:
+		return ipsec.NewStrongSwan(localSubnets, localEndpoint)
+
+	case WireGuard:
+		// TODO
+		err := fmt.Errorf("WireGuard is unsupported yet")
+		return nil, err
+	default:
+		// TODO define ERROR
+		return nil, fmt.Errorf("Unsupported backend type - %s", localEndpoint.Spec.Backend)
+	}
+
+
 }
